bnxt_re/lib: Fix for larger WR post_list sizes
In bnxt_re_send_wr_complete(), there is a type truncation bug.

The total count of slots saved in cur_slots_cnt is a uint32_t.
This gets assigned to a uint8_t (slots) before passing it to
bnxt_re_incr_tail() since it takes a uint8_t argument for the
slot count. And in the process, the value gets truncated.

Fix the type of slots variable.

Fixes: 6514d99eb8aa ("bnxt_re/lib: Add WR API send functions")

+++ providers/bnxt_re/verbs.c
@@ -1721,7 +1721,7 @@ static int bnxt_re_send_wr_complete(struct ibv_qp_ex *ibvqp)
 	struct bnxt_re_qp *qp = to_bnxt_re_qp((struct ibv_qp *)ibvqp);
 	struct bnxt_re_queue *sq = qp->jsqq->hwque;
 	int err = qp->wr_sq.error;
-	uint8_t slots;
+	uint32_t slots;
 
 	if (unlikely(err))
 		goto exit;
